+++ src/fontra/backends/base.py
@@ -2,6 +2,7 @@ from typing import Any
 
 from ..core.classes import (
     Axes,
+    ConditionalSubstitutions,
     FontInfo,
     FontSource,
     Kerning,
@@ -53,6 +54,9 @@ class ReadableBaseBackend:
     async def getGlyphInfos(self) -> dict[str, Any]:
         return {}
 
+    async def getConditionalSubstitutions(self) -> ConditionalSubstitutions:
+        return ConditionalSubstitutions()
+
 
 class WritableBaseBackend(ReadableBaseBackend):
     async def putGlyph(
@@ -89,3 +93,8 @@ class WritableBaseBackend(ReadableBaseBackend):
 
     async def putGlyphInfos(self, glyphInfos: dict[str, Any]) -> None:
         pass  # Better drop the glyph infos than to crash
+
+    async def putConditionalSubstitutions(
+        self, substitutions: ConditionalSubstitutions
+    ) -> None:
+        raise NotImplementedError()